pyproject.toml
[project]
name = "pychangelog"
license = "AGPL-3.0"
version = "0.2.0"
description = "A Python script that can generate a changelog from your Github issues and pull requests, filtering and organizing them by tag."
authors = [
    { name = "RERO+", email = "software@rero.ch"}
]

requires-python = ">=3.9"

dependencies = [
    "PyGithub>=1.55",
    "click>=8.1.2",
]

[dependency-groups]
dev = [
    "poethepoet>=0.36.0",
    "ruff>=0.12.4",
]

[tool.ruff.lint]
extend-select = ["ALL"]
ignore = ["ANN", "C901", "COM812", "D100", "DTZ", "PGH004", "PLR0912", "PTH"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.poe.tasks]
changelog = {cmd = "./pychangelog/changelog.py", help = "Generate changelog"}

[build-system]
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["."]

[tool.uv]
package = true
